Still working on integrating imgspect.

+++ rails3/app/models/subregion.rb
@@ -5,8 +5,8 @@ class Subregion < SparqlModel
     @endpoint = Rails.configuration.sparql_endpoint
     @attributes = {
       :cite_urn => [ "this:cite_urn", ::URN, SINGLE, REQUIRED ],
-      :x => [ "this:x", ::Integer, SINGLE, REQUIRED ],
-      :y => [ "this:y", ::Integer, SINGLE, REQUIRED ],
+      :x => [ "this:x", ::Float, SINGLE, REQUIRED ],
+      :y => [ "this:y", ::Float, SINGLE, REQUIRED ],
       :width => [ "this:width", ::Float, SINGLE, REQUIRED ],
       :height => [ "this:height", ::Float, SINGLE, REQUIRED ],
       :caption => [ "this:caption", ::String, SINGLE ]
